added check for no csharp extension installed
